rts-ml-inference-service/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy~=2.1.3
pandas~=2.3.1
scikit-learn~=1.7.1


psycopg2-binary~=2.9.10
keras~=3.10.0
Flask~=3.1.1